backend/requirements.txt
Django
djangorestframework
spacy
uvicorn
gunicorn
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
django-cors-headers
